Move Paginator to own module
